[INDY-1134] Move test to indy_node, created confteset and moved fixtures
